9 lab

+++ app/Http/Controllers/ArticleController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Article;
+use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
+
 
 class ArticleController extends Controller
 {
@@ -13,7 +15,7 @@ class ArticleController extends Controller
      */
     public function index()
     {
-       $articles = Article::latest()->paginate(5);
+        $articles = Article::latest()->paginate(5);
         return view('/article/article', ['articles'=>$articles]);
     }
 
@@ -22,9 +24,8 @@ class ArticleController extends Controller
      */
     public function create()
     {
-        //
-		Gate::authorize('create', Article::class);
-		return view('/article/create');
+        Gate::authorize('create', Article::class);
+        return view('article.create');
     }
 
     /**
@@ -32,20 +33,19 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Article::class);
         $request->validate([
-			'date'=> 'required|date',
-			'title'=> 'required|min:10',
-			'text'=>'max:100'
-		]);
-
-		$article = new Article;
-		$article->date_public = $request->date;
-		$article->title = request('title');
-		$article->text = $request->text;
-		$article->users_id = auth()->id();
-		$article->save();
-
-		return redirect()->route('article.index')->with('message','Create successful');
+            'date' => 'required|date',
+            'title' => 'required|min:10',
+            'text' => 'max:100'
+        ]);
+        $article = new Article;
+        $article->date_public = $request->date;
+        $article->title = request('title');
+        $article->text = $request->text;
+        $article->users_id = auth()->id();
+        $article->save();
+        return redirect()->route('article.index')->with('message','Create successful');
     }
 
     /**
@@ -53,9 +53,8 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        $article->load('comments');
-    
-    return view('article/show', ['article' => $article]);
+        $comments = Comment::where('article_id', $article->id)->get();
+        return view('article.show', ['article'=>$article, 'comments'=>$comments]);
     }
 
     /**
@@ -63,8 +62,8 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
-		 return view('article.edit', ['article'=>$article]);
-        
+        Gate::authorize('restore', $article);
+        return view('article.edit', ['article'=>$article]);
     }
 
     /**
@@ -72,6 +71,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('update', $article);
         $request->validate([
             'date' => 'required|date',
             'title' => 'required|min:10',
@@ -90,7 +90,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-         $article->delete();
+        Gate::authorize('delete', $article);
+        $article->delete();
         return redirect()->route('article.index')->with('message','Delete successful');
     }
 }
